--- a/src/ess/dream/io/geant4.py
+++ b/src/ess/dream/io/geant4.py
@@ -167,15 +167,11 @@ def _extract_detector(
     return events
 
 
-def get_source_position(
-    raw_source: RawSource[RunType],
-) -> SourcePosition[RunType]:
+def get_source_position(raw_source: RawSource[RunType]) -> SourcePosition[RunType]:
     return SourcePosition[RunType](raw_source["position"])
 
 
-def get_sample_position(
-    raw_sample: RawSample[RunType],
-) -> SamplePosition[RunType]:
+def get_sample_position(raw_sample: RawSample[RunType]) -> SamplePosition[RunType]:
     return SamplePosition[RunType](raw_sample["position"])
 
 
@@ -184,10 +180,11 @@ def patch_detector_data(
     source_position: SourcePosition[RunType],
     sample_position: SamplePosition[RunType],
 ) -> ReducibleDetectorData[RunType]:
-    out = detector_data.copy(deep=False)
-    out.coords["source_position"] = source_position
-    out.coords["sample_position"] = sample_position
-    return ReducibleDetectorData[RunType](out)
+    return ReducibleDetectorData[RunType](
+        detector_data.assign_coords(
+            source_position=source_position, sample_position=sample_position
+        )
+    )
 
 
 def geant4_detector_dimensions(