--- a/src/ess/powder/grouping.py
+++ b/src/ess/powder/grouping.py
@@ -17,7 +17,7 @@ from .types import (
 def _drop_grouping_and_bin(
     data: sc.DataArray, *, dims_to_reduce: tuple[str, ...] | None = None, edges: dict
 ) -> sc.DataArray:
-    all_pixels = data.bins.concat(dims_to_reduce)
+    all_pixels = data if dims_to_reduce == () else data.bins.concat(dims_to_reduce)
     # all_pixels may just have a single bin now, which currently yields
     # inferior performance when binning (no/bad multi-threading?).
     # We operate on the content buffer for better multi-threaded performance.
@@ -45,10 +45,10 @@ def focus_data_dspacing_and_two_theta(
     dspacing_bins: DspacingBins,
     twotheta_bins: TwoThetaBins,
 ) -> FocussedDataDspacingTwoTheta[RunType]:
-    edges = {twotheta_bins.dim: twotheta_bins, dspacing_bins.dim: dspacing_bins}
-    dims_to_reduce = None if 'two_theta' in data.bins.coords else ()
     return FocussedDataDspacingTwoTheta[RunType](
-        _drop_grouping_and_bin(data, edges=edges, dims_to_reduce=dims_to_reduce)
+        data.bin({twotheta_bins.dim: twotheta_bins}).bin(
+            {dspacing_bins.dim: dspacing_bins}
+        )
     )