--- a/tests/dream/io/geant4_test.py
+++ b/tests/dream/io/geant4_test.py
@@ -113,7 +113,7 @@ def test_load_geant4_csv_mantle_has_expected_coords(file):
     assert "sector" not in mantle.bins.coords
     assert "tof" in mantle.bins.coords
     assert "wavelength" in mantle.bins.coords
-    assert "position" in mantle.bins.coords
+    assert "position" in mantle.coords
 
 
 def test_load_geant4_csv_endcap_backward_has_expected_coords(file):
@@ -128,7 +128,7 @@ def test_load_geant4_csv_endcap_backward_has_expected_coords(file):
     assert "sector" not in endcap.bins.coords
     assert "tof" in endcap.bins.coords
     assert "wavelength" in endcap.bins.coords
-    assert "position" in endcap.bins.coords
+    assert "position" in endcap.coords
 
 
 def test_load_geant4_csv_endcap_forward_has_expected_coords(file):
@@ -143,7 +143,7 @@ def test_load_geant4_csv_endcap_forward_has_expected_coords(file):
     assert "sector" not in endcap.bins.coords
     assert "tof" in endcap.bins.coords
     assert "wavelength" in endcap.bins.coords
-    assert "position" in endcap.bins.coords
+    assert "position" in endcap.coords
 
 
 def test_load_geant4_csv_high_resolution_has_expected_coords(file):
@@ -157,7 +157,7 @@ def test_load_geant4_csv_high_resolution_has_expected_coords(file):
 
     assert "tof" in hr.bins.coords
     assert "wavelength" in hr.bins.coords
-    assert "position" in hr.bins.coords
+    assert "position" in hr.coords
 
 
 def test_load_geant4_csv_sans_has_expected_coords(file):
@@ -174,7 +174,7 @@ def test_load_geant4_csv_sans_has_expected_coords(file):
 
     assert "tof" in sans.bins.coords
     assert "wavelength" in sans.bins.coords
-    assert "position" in sans.bins.coords
+    assert "position" in sans.coords
 
 
 def test_geant4_in_pipeline(file_path, file):