### Changes

- Bypass the same single-threading problem encountered in SANS, I think we either have to put a helper into ESSreduce, or see if the underlying problem should be fixed in Scipp.
- Avoid storing position for every event. This will only affect Geant4 data, so the improvements from that will not be seen for NeXus data. Note that this can lead to NaN positions if there are no events. Is this a problem (not in the tests)? If so we either need to revert, or find a better solution.
- Fixes #41 (but I guess it has to be revisited sooner or later). I did not spot any other "obvious" problems in the functions taking the most time.

### Baseline

Baseline for Geant4 workflow:

```
2.47 seconds load_geant4_csv
4.75 seconds load_geant4_csv
3.00 seconds extract_geant4_detector_data
6.11 seconds extract_geant4_detector_data
0.62 seconds normalize_by_proton_charge
4.38 seconds add_scattering_coordinates_from_positions
0.35 seconds apply_masks
0.99 seconds normalize_by_proton_charge
4.36 seconds convert_to_dspacing
5.20 seconds add_scattering_coordinates_from_positions
0.23 seconds apply_masks
2.94 seconds convert_to_dspacing
9.75 seconds focus_data_dspacing
10.12 seconds focus_data_dspacing
1.19 seconds normalize_by_vanadium_dspacing
```

I artificially increased the data size by modifying the data after loading:

```python
def extract_geant4_detector_data(
    detector: RawDetector[RunType],
) -> RawDetectorData[RunType]:
    """Extract the histogram or event data from a loaded GEANT4 detector."""
    data = extract_detector_data(detector)
    data = data.bins.concatenate(data)
    data = data.bins.concatenate(data)
    data = data.bins.concatenate(data)
    data = data.bins.concatenate(data)
    data = data.bins.concatenate(data)
    data = data.bins.concatenate(data) 
    data = data.bins.concatenate(data)  # ~ 10 GByte (sample run)
    return RawDetectorData[RunType](data)
```

The timings reported for this function are thus irrelevant in reality.

### This branch

`RawDetectorData[SampleRun]` is now 6 GByte instead of 10 GByte:
```
2.45 seconds load_geant4_csv
5.28 seconds load_geant4_csv
1.97 seconds extract_geant4_detector_data
4.03 seconds extract_geant4_detector_data
0.34 seconds normalize_by_proton_charge
0.20 seconds convert_to_dspacing
2.14 seconds focus_data_dspacing
0.46 seconds normalize_by_proton_charge
0.23 seconds apply_masks
0.20 seconds convert_to_dspacing
2.14 seconds focus_data_dspacing
0.94 seconds normalize_by_vanadium_dspacing
```