@SimonHeybrock 2024-06-25
Converted to draft, since it looks like I messed up memory consumption (at least CI is having trouble, it seems).