@SimonHeybrock 2024-06-25
Looking at https://github.com/scipp/essdiffraction/actions/workflows/ci.yml it seems the CI timings were a perfectly normal outlier. Ready for review once more!